+++ backend/src/main/java/sw_css/member/api/MemberController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import sw_css.member.application.MemberCommandService;
 import sw_css.member.application.MemberQueryService;
 import sw_css.member.application.dto.request.ChangePasswordRequest;
 import sw_css.member.application.dto.response.StudentMemberResponse;
@@ -24,6 +25,7 @@ import sw_css.utils.annotation.MemberInterface;
 @Transactional
 public class MemberController {
     private final MemberQueryService memberQueryService;
+    private final MemberCommandService memberCommandService;
 
     @GetMapping("/{memberId}")
     public ResponseEntity<StudentMemberResponse> findStudent(@PathVariable final Long memberId) {
@@ -33,7 +35,7 @@ public class MemberController {
     @PatchMapping("/change-password")
     public ResponseEntity<Void> changeMemberPassword(@MemberInterface Member me,
                                                      @RequestBody @Valid ChangePasswordRequest request) {
-        memberQueryService.changePassword(me, request.oldPassword(), request.newPassword());
+        memberCommandService.changePassword(me, request.oldPassword(), request.newPassword());
         return ResponseEntity.noContent().build();
     }
 }
